Extract AudioFileProcessorWaveView
Extract the class `AudioFileProcessorWaveView` into its own files.
